BoardUpdate1.8
Interactions with swapping pieces are now working properly. Is there will be no match -> switch back.
If there would be a match after switch -> deactivte pieces from secne (for now). Still kinda glitchy and needs to be worked on.

Also fixed board initialization to not have any matches on start, should be working 100%